requirements.txt
streamlit
boto3
pandas
openpyxl
python-docx
PyPDF2
tiktoken